requirements.txt
streamlit>=1.45.0
langchain>=1.0.0
langchain-google-genai>=2.1.3
langchain-openai>=0.3.0
langchain-core>=0.3.41
langgraph>=0.3.0
requests>=2.32.3
pydantic>=2.11.1
python-dotenv>=1.1.0
pandas>=2.0.0
